Move Unshielding dialog to AccountDetailsFragment
