+++ app/src/main/res/layout/activity_unshielding_accounts.xml
@@ -56,7 +56,7 @@
 
     <Button
         android:id="@+id/done_button"
-        style="@style/CCX_Button_Primary"
+        style="@style/MW24_Button_Primary"
         android:layout_width="match_parent"
         android:layout_marginHorizontal="18dp"
         android:layout_marginBottom="20dp"
